Hell, it's about time to refactor Sonderbot
